+++ lib/cocoapods/installer/analyzer.rb
@@ -360,7 +360,7 @@ module Pod
         check_pod_target_swift_versions(specs_by_target)
         pod_targets = generate_pod_targets(specs_by_target)
         aggregate_targets = specs_by_target.keys.map do |target_definition|
-          generate_target(target_definition, pod_targets)
+          generate_aggregate_target(target_definition, pod_targets)
         end
         if installation_options.integrate_targets?
           # Copy embedded target pods that cannot have their pods embedded as frameworks to
@@ -397,7 +397,7 @@ module Pod
       #
       # @return [AggregateTarget]
       #
-      def generate_target(target_definition, pod_targets)
+      def generate_aggregate_target(target_definition, pod_targets)
         target = AggregateTarget.new(target_definition, sandbox)
         target.host_requires_frameworks |= target_definition.uses_frameworks?
 
@@ -419,10 +419,28 @@ module Pod
         end
 
         target.pod_targets = pod_targets.select do |pod_target|
-          pod_target.target_definitions.include?(target_definition)
+          pod_target.target_definitions.include?(target_definition) &&
+              (target_definition.non_inherited_dependencies.map(&:name).include?(pod_target.name) || !pod_target_test_only?(pod_target, pod_targets))
         end
 
         target
+      end
+
+      # Returns true if a pod target is only used by other pod targets as a test dependency and therefore should
+      # not be included as part of the aggregate target.
+      #
+      # @param [PodTarget] pod_target
+      #        the pod target being queried
+      #
+      # @param [Array<PodTarget>] pod_targets
+      #        the array of pod targets to check against
+      #
+      # @return [Boolean] if the pod target is only referenced from test dependencies.
+      #
+      def pod_target_test_only?(pod_target, pod_targets)
+        source = pod_targets.any? { |pt| pt.dependent_targets.map(&:name).include?(pod_target.name) }
+        test = pod_targets.any? { |pt| pt.test_dependent_targets.map(&:name).include?(pod_target.name) }
+        !source && test
       end
 
       # Verify that targets using a pod have the same swift version
@@ -496,8 +514,6 @@ module Pod
             test_dependencies = transitive_dependencies_for_specs(target.specs.select(&:test_specification?), target.platform, all_specs).group_by(&:root)
             target.dependent_targets = filter_dependencies(dependencies, pod_targets_by_name, target)
             target.test_dependent_targets = filter_dependencies(test_dependencies, pod_targets_by_name, target)
-            # Test dependendent targets include our own target as a test dependency.
-            target.test_dependent_targets << target
           end
         else
           dedupe_cache = {}
@@ -512,8 +528,6 @@ module Pod
               test_dependencies = transitive_dependencies_for_specs(target.specs.select(&:test_specification?), target.platform, all_specs).group_by(&:root)
               target.dependent_targets = pod_targets.reject { |t| dependencies[t.root_spec].nil? }
               target.test_dependent_targets = pod_targets.reject { |t| test_dependencies[t.root_spec].nil? }
-              # Test dependendent targets include our own target as a test dependency.
-              target.test_dependent_targets << target
             end
           end
         end